Renamed files and deleted TCP versions
